Move login/logout functionality to the 'sessions' controller
