+++ config/routes.rb
@@ -1,9 +1,9 @@
 Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
   root 'welcome#index'
-  get '/login', to: 'welcome#login_form', as: 'login'
-  post '/login', to: 'welcome#login'
-  post '/logout', to: 'welcome#logout', as: 'logout'
+  get '/login', to: 'sessions#login_form', as: 'login'
+  post '/login', to: 'sessions#login'
+  post '/logout', to: 'sessions#logout', as: 'logout'
 
   # Build the category routes for albums, books and movies
   category_constraints = { category: /(albums)|(books)|(movies)/}
